Bugfix: Use sane Markdown renderer (Redcarpet)
IDK, the other thing just kept trying to encode its regexes
in the Windows-31J string encoding and then being mad that
it was being asked to process UTF-8 text.  When I tried to
convert the UTF-8 string to said bonkers format, Ruby said
that it couldn't do it.  So, whatever, we'll just use a
parser that isn't confused by character encodings in the
simplest use case imaginable.

Added benefit: We can just use their code for auto-
generating links, instead of writing out own regular
expressions to do it on the fly.